UPSTREAM: fs-verity: rename fsverity_signed_digest to fsverity_formatted_digest
The name "struct fsverity_signed_digest" is causing confusion because it
isn't actually a signed digest, but rather it's the way that the digest
is formatted in order to be signed.  Rename it to
"struct fsverity_formatted_digest" to prevent this confusion.

Also update the struct's comment to clarify that it's specific to the
built-in signature verification support and isn't a requirement for all
fs-verity users.

I'll be renaming this struct in fsverity-utils too.

Acked-by: Luca Boccassi <luca.boccassi@microsoft.com>
Link: https://lore.kernel.org/r/20201113211918.71883-3-ebiggers@kernel.org
Signed-off-by: Eric Biggers <ebiggers@google.com>
Bug: 186190934
Test: With rest of series and ABI update incrementalfs.ko loads
(cherry picked from commit 9e90f30e78572ecfc1c74c735a034c955d822ba6)
Signed-off-by: Paul Lawrence <paullawrence@google.com>
Change-Id: I70ed689387efb9431418e7f71c3f443980084be1